+++ src/components/CategoryLabel.tsx
